api_comparison/requirements.txt
openai>=1.82.0
python-dotenv>=1.1.0
azure-identity>=1.21.0
rich>=14.0.0
requests>=2.32.0
yfinance>=0.2.40
